Ci/cd & db transition changes

+++ src/main/java/com/crumbs/fss/service/RestaurantService.java
@@ -1,33 +1,35 @@
 package com.crumbs.fss.service;
-
 import com.crumbs.fss.DTO.addRestaurantDTO;
 import com.crumbs.fss.DTO.updateRestaurantDTO;
-import com.crumbs.fss.ExceptionHandling.DuplicateEmailException;
 import com.crumbs.fss.ExceptionHandling.DuplicateLocationException;
-import com.crumbs.fss.entity.*;
-import com.crumbs.fss.entity.MenuItem;
-import com.crumbs.fss.repository.*;
+import com.crumbs.lib.entity.*;
+import com.crumbs.lib.entity.MenuItem;
+import com.crumbs.lib.repository.*;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cglib.core.CollectionUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-
 import javax.persistence.EntityNotFoundException;
-import java.awt.*;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Optional;
 
 @Service
 @Transactional(rollbackFor = { Exception.class })
 public class RestaurantService {
 
-    @Autowired RestaurantRepository restaurantRepository;
-    @Autowired MenuItemRepository menuItemRepository;
-    @Autowired LocationRepository locationRepository;
-    @Autowired RestaurantCategoryRepository restaurantCategoryRepository;
-    @Autowired RestaurantOwnerRepository restaurantOwnerRepository;
-    @Autowired UserDetailRepository userDetailRepository;
+    @Autowired
+    RestaurantRepository restaurantRepository;
+    @Autowired
+    MenuItemRepository menuItemRepository;
+    @Autowired
+    LocationRepository locationRepository;
+    @Autowired
+    RestaurantCategoryRepository restaurantCategoryRepository;
+    @Autowired
+    RestaurantOwnerRepository restaurantOwnerRepository;
+    @Autowired UserDetailsRepository userDetailRepository;
+    @Autowired RestaurantStatusRepository restaurantStatusRepository;
+    @Autowired
+     UserStatusRepository userStatusRepository;
 
     public List<Restaurant> getAllRestaurants(){
         return restaurantRepository.findAll();
@@ -39,13 +41,13 @@ public class RestaurantService {
 
     public Restaurant addRestaurant(addRestaurantDTO a) {
 
-         if(userDetailRepository.findUserByEmail(a.getEmail())!=null)
-            throw new DuplicateEmailException();
+//         if(userDetailRepository.findUserByEmail(a.getEmail())!=null)
+//            throw new DuplicateEmailException();
 
         if(locationRepository.findLocationByStreet(a.getStreet())!=null)
             throw new DuplicateLocationException();
 
-        UserDetail userDetail = UserDetail.builder()
+        UserDetails userDetail = UserDetails.builder()
                 .firstName(a.getFirstName())
                 .lastName(a.getLastName())
                 .email(a.getEmail())
@@ -53,8 +55,8 @@ public class RestaurantService {
 
         userDetailRepository.save(userDetail);
 
-        RestaurantOwner restaurantOwner = RestaurantOwner.builder()
-                .userDetail(userDetail)
+        Owner restaurantOwner = Owner.builder()
+                .userDetails(userDetail)
                 .build();
 
         restaurantOwnerRepository.save(restaurantOwner);
@@ -79,7 +81,7 @@ public class RestaurantService {
 
         List<String> categories = a.getCategories();
         if(categories!= null && !categories.isEmpty()) {
-            categories.forEach(category -> restaurantCategoryRepository.insertRestaurantCategory(category, restaurant.getId()));
+            categories.forEach(category -> restaurantCategoryRepository.insertRestaurantCategory(restaurant.getId(), category));
         }
         return restaurant;
     }
@@ -98,8 +100,8 @@ public class RestaurantService {
 
         Restaurant temp = restaurantRepository.findById(id).orElseThrow(EntityNotFoundException::new);
 
-        if(updateRestaurantDTO.getEmail() != null && !updateRestaurantDTO.getEmail().equals(temp.getRestaurantOwner().getUserDetail().getEmail()) && userDetailRepository.findUserByEmail(updateRestaurantDTO.getEmail())!=null)
-            throw new DuplicateEmailException();
+//        if(updateRestaurantDTO.getEmail() != null && !updateRestaurantDTO.getEmail().equals(temp.getRestaurantOwner().getUserDetails().getEmail()) && userDetailRepository.findUserByEmail(updateRestaurantDTO.getEmail())!=null)
+//            throw new DuplicateEmailException();
 
         if(updateRestaurantDTO.getStreet() != null && !updateRestaurantDTO.getStreet().equals(temp.getLocation().getStreet()) && locationRepository.findLocationByStreet(updateRestaurantDTO.getStreet())!=null)
             throw new DuplicateLocationException();
@@ -107,15 +109,15 @@ public class RestaurantService {
         // Update User Details
         String firstName = updateRestaurantDTO.getFirstName();
         if(firstName!= null && !firstName.isEmpty())
-            temp.getRestaurantOwner().getUserDetail().setFirstName(firstName);
+            temp.getRestaurantOwner().getUserDetails().setFirstName(firstName);
 
         String lastName = updateRestaurantDTO.getLastName();
         if(firstName!= null && !firstName.isEmpty())
-            temp.getRestaurantOwner().getUserDetail().setLastName(lastName);
+            temp.getRestaurantOwner().getUserDetails().setLastName(lastName);
 
         String email = updateRestaurantDTO.getEmail();
         if(email!= null && !email.isEmpty())
-            temp.getRestaurantOwner().getUserDetail().setEmail(email);
+            temp.getRestaurantOwner().getUserDetails().setEmail(email);
 
         //Update Restaurant Location
         String street = updateRestaurantDTO.getStreet();
@@ -126,9 +128,9 @@ public class RestaurantService {
         if(city!= null && !city.isEmpty())
             temp.getLocation().setCity(city);
 
-        Integer zip = updateRestaurantDTO.getZip();
+        String zip = updateRestaurantDTO.getZip();
         if(zip!= null)
-            temp.getLocation().setZipCode(zip);
+           temp.getLocation().setZipCode(zip);
 
         String state = updateRestaurantDTO.getState();
         if(state!= null && !state.isEmpty())
@@ -150,29 +152,39 @@ public class RestaurantService {
         //replace with new ones
         List<String> newCategories = updateRestaurantDTO.getCategories();
         if(newCategories!= null && !newCategories.isEmpty()) {
-            newCategories.forEach(category -> {
-            restaurantCategoryRepository.insertRestaurantCategory(category,temp.getId());
-            });
+            newCategories.forEach(category -> restaurantCategoryRepository.insertRestaurantCategory(temp.getId(), category));
         }
 
-//        //update menu items
+        //update menu items
         List<MenuItem> oldMenu = temp.getMenuItems();
         List<MenuItem> newMenu = updateRestaurantDTO.getMenu();
-//
-        newMenu.forEach(item -> {
-            if(item.getId() == null)
-                item.setRestaurant(temp);
 
-        });
-        oldMenu.forEach( item ->{
-            if(!newMenu.contains(item))
-                menuItemRepository.delete(item);
-        });
+        //only update menu if there were changes
+        if(newMenu != null && !newMenu.isEmpty()) {
 
-        temp.setMenuItems(newMenu);
+            //for newly added menu items, set restaurant to restaurant (otherwise restaurant_ID stays null on save)
+            newMenu.forEach(item -> {
+                if (item.getId() == null)
+                    item.setRestaurant(temp);
+            });
+            temp.setMenuItems(newMenu);
 
+            //for old menu items that are now deleted, delete them (they don't automatically delete on save)
+            if(oldMenu != null && !oldMenu.isEmpty()) {
+                oldMenu.forEach(item -> {
+                    if (!newMenu.contains(item))
+                        menuItemRepository.delete(item);
+                });
+            }
+        }
 
         return restaurantRepository.save(temp);
+    }
+    public void requestDeleteRestaurant(Long id){
+        Restaurant temp = restaurantRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+        RestaurantStatus status = restaurantStatusRepository.findById("PENDING_DELETE").get();
+        temp.setRestaurantStatus(status);
+        restaurantRepository.save(temp);
     }
